msvc14 updates

+++ examples/cci/ex22_Search_Predicate/ex22_simple_ip.h
@@ -47,7 +47,7 @@ public:
     uchar_param("uchar_param", 'U'),
     int_param("int_param", -1),
     uint_param("uint_param", 0xFEEDF00D),
-    float_param("float_param", 3.14),
+    float_param("float_param", (float)3.14),
     double_param("double_param", 2.714),
     string_param("string_param", "C++ String") {
         XREPORT("@Ctor: Default value of " << bool_param.name()
